requirements.txt
diffusers>=0.36.0
transformers>=4.51.3
peft>=0.17.1
accelerate>=1.2.1
safetensors
torchmetrics>=1.4
timm>=1.0
opencv-python
imagehash
einops
rich>=13.9.4
tqdm
scikit-learn
pandas
matplotlib
